Move module level headers next to their modules
(Rather than inside their modules)
