+++ inexor/test/net/DuplexMCTest.h
@@ -3,7 +3,7 @@
 
 #include "gtest/gtest.h"
 #include "inexor/test/net/SimplexMCTest.h"
-#include "inexor/net/net.h"
+#include "inexor/net.h"
 
 template<typename T_other>
 class DuplexMCTest_SwapProvider : public SimplexMCTest_Provider {
