+++ inexor/util.h
